feat(secret): add support for custom config via environment variable
This commit introduces the ability for the `SecretStore` class to read
its configuration from an environment variable, specifically
`KEEPASSXC_RUN_GCKPXC_CONFIG`. If this variable is set, its value will
be used as a flag for "gir-credential-keepassxc" command.
This enhancement prevents overwrite a local configuration for "gir-credential-keepassxc".

Additionally, a corresponding test fixture has been added to ensure that
the environment variable is correctly set during tests, enabling
verification of the new functionality. This improves the testing
coverage and ensures that the feature works as intended.